[untested] mega raid barriers only heal some fraction of the maxHP
instead of maxHP when broken, mega raid barriers now only a part of the maxHP.
heals maxHP*(remaining barriers + 1)/(total barriers + 1)

example for 3 shields (max value): heal 3/4 of the total health, then 2/4, and lastly 1/4.

idea is to combine this with the heal shockwave.
maybe the actual heal values could depend on the raid rank as well