[1/2] port Marlux's IV viewing in conditions menu
port the hexa graph in the condition submenu of the pokenav and made it to view IVs. merely dummied the part where real condition data is read so I can switch easily between the two. next step is to be able to view both (and EVs?) by pressing select (like merrp did).