huge commit, see desc
made some space in primary tileset. gonna remove the puddle towards secondary, still to find the last usage of it.
ported a few water edge tiles (bootom mostly). also added the grass cover for cliff edges, but didn't make the metatiles for it
Fixed a shitload of maps with my tile changes